fix(bowerCheck): don't ask the user what version works best, always use latest
When Bower asks the user which version it needs to install, it's a question which is not the user's problem. It's unnecessary, gives headaches and should be removed. The goal of Angus is to give you a smooth experience, and questions like these mess that up. Read http://www.joelonsoftware.com/uibook/chapters/fog0000000059.html
Strict versioning isn't as important for bower as it is for npm, which deals with a lot more critical things like disk latency, networking and most of all security. If something on the client doesn't work, you will notice it when you open up the browser. Use the strict versioning syntax <package>@<version> in angus.config.js if you're paranoid.